Overlays stay in kml.py
